Attempt to deduplicate more
